% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSimModel.R
\name{buildSimModel}
\alias{buildSimModel}
\title{Builds a lavaan model}
\usage{
buildSimModel(EScond = 0.2, ESmod = 0.2, ESint = 0.2,
  bpath = c(0.4, 0.3), model = "growth")
}
\arguments{
\item{EScond}{effect size condition}

\item{ESmod}{effect size moderator}

\item{ESint}{effect size interaction term}

\item{bpath}{vector of regression coefficients from mediator to latent intercept and slope respectively}

\item{model}{latent growth model (model = "growth") with time measurements is build or moderated mediation model with four dependent variables}
}
\value{
a lavaan model
}
\description{
This function builds a latent growth model (model = "growth") or a moderated mediation model with four dependent variables
(model = NULL) to be used in lavaan. Function is called by simPwr.growth().
}
\examples{
buildModel()
}
\keyword{growth}
\keyword{interaction}
\keyword{latent}
\keyword{lavaan}
\keyword{mediation}
\keyword{moderation}
