% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawSamples.R
\name{drawSamples}
\alias{drawSamples}
\title{Simulation for power of moderation model}
\usage{
drawSamples(preds = NULL, Sigma = NULL, coef = NULL,
  predictNames = NULL, dependName = NULL, interactionTerms = NULL,
  repetitions = 100, samples = 1000)
}
\arguments{
\item{preds}{The number of predictor variables}

\item{Sigma}{The correlation matrix of the predictor variables}

\item{coef}{The regression coefficients of respectively the predictors and (if present) the interaction terms}

\item{predictNames}{Names of the predictors (usually left NULL, by default called "x1", "x2" etc.)}

\item{dependName}{Name of the dependent variable (usually left NULL, by default called "y")}

\item{interactionTerms}{List of vectors of length two, that specifies the two-way interaction terms, e.g. interactionTerms = list(c("x1","x2"), c("x1","x3"))}

\item{repetitions}{The number of records (sample size) in a single data frame}

\item{samples}{The number of samples drawn from the specified model (replications)}
}
\value{
A matrix with in the rows representing the samples, and columns the R-squared values and the p-values, respectively
}
\description{
This function allows you to estimate the power in an linear model with interaction effects, used by SimPwr.regr().
}
\details{
Instead of interaction terms also quadratic terms can be specified similar to interaction terms, e.g. interactionTerms = list(c("x1","x1"))
}
\examples{
drawSamples(preds = 3, coef = c(.4,.3.,2), repetitions = 100, samples=100)
}
\keyword{interaction}
\keyword{moderation}
\keyword{power}
\keyword{regression}
