% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power regression.R
\name{print.simPower.moderation}
\alias{print.simPower.moderation}
\title{Print method for simPower.moderation}
\usage{
\method{print}{simPower.moderation}(x, ...)
}
\arguments{
\item{x}{an object of class simPower.moderation}
}
\description{
This function allows you to plot the power of an object of class simPower.moderation
}
\examples{
print(x)
}
\keyword{print}
\keyword{simPower.moderation}
