% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPwr.mod.R
\name{print.simPwr.mod}
\alias{print.simPwr.mod}
\title{Print method for simPwr.mod}
\usage{
\method{print}{simPwr.mod}(x, ...)
}
\arguments{
\item{x}{an object of class simPwr.mod}
}
\description{
This function allows you to plot the power of an object of class simPower.moderation
}
\examples{
print(x)
}
\keyword{print}
\keyword{simPwr.mod}
