% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPwr.regr.R
\name{print.simPwr.regr}
\alias{print.simPwr.regr}
\title{Print method for function regrPwrSim}
\usage{
\method{print}{simPwr.regr}(x, digits = x$input$digits, ...)
}
\arguments{
\item{x}{simPwr.regr object}
}
\description{
This function prints the results of a power analyses of the linear model
}
\examples{
print(x)
}
