% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sim_rdd}
\alias{sim_rdd}
\title{Simulate toy data for RDD examples}
\usage{
sim_rdd(n = 2000, jump = 0.3, sd_e = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{jump}{True discontinuity at the cutoff.}

\item{sd_e}{Noise SD.}
}
\value{
A data frame with columns \code{x} (running variable, cutoff at 0), \code{y}.
}
\description{
Simulate toy data for RDD examples
}
