% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPwr.lgm.R
\name{simPwr.lgm}
\alias{simPwr.lgm}
\title{Simulation for power of latent growth model}
\usage{
simPwr.lgm(samSizes, esSizes, ndepend = 4, ESmod = 0.3, ESint = 0.2,
  bpath = c(0.5, -0.1), rho = c(0, 0.1), error = 1, alpha = 0.05,
  maxiter = 1000)
}
\arguments{
\item{samSizes}{vector of sample size}

\item{esSizes}{vector of effect sizes of the condition effect}

\item{ndepend}{number of dependent variables}

\item{ESmod}{effect size moderator}

\item{ESint}{effect size interaction term}

\item{bpath}{vector of regression coefficients from mediator to latent intercept and slope respectively}

\item{rho}{vector with minimum and maximum value of randomly selected corrrelation between the variables}

\item{error}{standard deviations of the random error added to the data}

\item{alpha}{alpha level}

\item{niter}{number of iterations}
}
\value{
List with the following elements

power: estimate for all effects

bias: estimate for all effects

raw: the raw results of the simulation

input: the input paramters used in the simulation
}
\description{
This function computes the power curves
of a latent growth model with k measurements, one mediation effect and one interaction effect.
}
\examples{
simPwr.growth()
}
\keyword{SEM}
\keyword{growth}
\keyword{latent}
\keyword{mediation}
