% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPwr.mmrm.R
\name{simPwr.mmrm}
\alias{simPwr.mmrm}
\title{Simulation for power of moderated mediation repeated measures (mmrm) model}
\usage{
simPwr.mmrm(n = 100, EScond = 0.2, ESmod = 0.2, ESint = 0.2,
  bpath = c(0.4, 0.3, 0.2, 0.1), ndepend = 4, rho = c(0, 0),
  error = 1, alpha = 0.05, maxiter = 100)
}
\arguments{
\item{n}{sample size}

\item{EScond}{effect size condition}

\item{ESmod}{effect size moderator}

\item{ESint}{effect size interaction term}

\item{bpath}{vector of regression coefficients from mediator to latent intercept and slope respectively}

\item{ndepend}{number of dependent variables}

\item{rho}{vector with minimum and maximum value of randomly selected corrrelation between the variables}

\item{error}{standard deviations of the random error added to the data}

\item{alpha}{alpha level}

\item{maxiter}{number of iterations}
}
\value{
List with the following elements

power: estimate for all effects

bias: estimate for all effects

raw: the raw results of the simulation

input: the input paramters used in the simulation
}
\description{
This function allows you to estimate the power for a given sample size 
of a repeated measures model with k measurements, one mediation effect and one interaction effect.
}
\examples{
simPwr.mmrm()
}
\keyword{SEM}
\keyword{growth}
\keyword{latent}
\keyword{mediation}
