% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.simpower}
\alias{summary.simpower}
\title{Summarise a simpower object}
\usage{
\method{summary}{simpower}(object, ...)
}
\arguments{
\item{object}{A \code{simpower} object.}

\item{...}{Unused.}
}
\description{
Summarise a simpower object
}
