% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPwr.regr.R
\name{simPwr.regr}
\alias{simPwr.regr}
\title{Simulation for power moderation for a given sample size}
\usage{
simPwr.regr(n = 100, predictors = NULL, cor = NULL, betas = NULL,
  interactions = NULL, sig.level = 0.05, maxiter = 1000,
  monteCarlo = FALSE)
}
\arguments{
\item{n}{The sample size}

\item{predictors}{A number of predictor variables}

\item{cor}{A vector or matrix containing the correlation bewtween the predictors, if NULL random correlations will be used}

\item{betas}{A vector regression coeffcients (effect sizes)}

\item{interactions}{A list of vectors containing a pair of variables}

\item{sig.level}{Type I error level for power analysis}

\item{maxiter}{The number of independent replications (iterations, samples)}

\item{monteCarlo}{Logical indicating whether function is used in a Monte Carlo experiment (TRUE)}
}
\description{
This function allows you to estimate the power in an linear model with interaction effects.
}
\examples{
simPwr.regr(n=100, predictors = 3, betas = c(.4,.3.,.2), maxiter= 1000)
}
\keyword{interaction}
\keyword{moderation}
\keyword{power}
\keyword{regression}
